+++ various/program-list.cpp
@@ -2,19 +2,16 @@
 /*
  *    Copyright (C) 2014
  *    Jan van Katwijk (J.vanKatwijk@gmail.com)
- *    Lazy Chair Programming
+ *    Lazy Chair Computing
  *
- *    This file is part of the  SDR-J series.
- *    Many of the ideas as implemented in the SDR-J are derived from
- *    other work, made available through the (a) GNU general Public License. 
- *    All copyrights of the original authors are recognized.
+ *    This file is part of the  drm2
  *
- *    SDR-J is free software; you can redistribute it and/or modify
+ *    drm2 is free software; you can redistribute it and/or modify
  *    it under the terms of the GNU General Public License as published by
  *    the Free Software Foundation; either version 2 of the License, or
  *    (at your option) any later version.
  *
- *    SDR-J is distributed in the hope that it will be useful,
+ *    drm2 is distributed in the hope that it will be useful,
  *    but WITHOUT ANY WARRANTY; without even the implied warranty of
  *    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
  *    GNU General Public License for more details.
@@ -42,8 +39,8 @@
 	         this, SLOT (tableSelect (int, int)));
 	connect (tableWidget, SIGNAL (cellDoubleClicked (int, int)),
 	         this, SLOT (removeRow (int, int)));
-	connect (this, SIGNAL (newFrequency (quint64)),
-	         mr, SLOT (setFrequency (quint64)));
+	connect (this, SIGNAL (newFrequency (int32_t)),
+	         mr, SLOT (setFrequency (int32_t)));
 	loadTable ();
 }
